+++ utils/src/commonMain/kotlin/com/badoo/reaktive/utils/Freeze.kt
@@ -3,3 +3,5 @@ package com.badoo.reaktive.utils
 expect fun <T> T.freeze(): T
 
 expect fun <T: Any> T.ensureNeverFrozen(): T
+
+expect val Any?.isFrozen: Boolean
