+++ utils/src/jvmJsCommonMain/kotlin/com/badoo/reaktive/utils/Freeze.kt
@@ -3,3 +3,5 @@ package com.badoo.reaktive.utils
 actual fun <T> T.freeze(): T = this
 
 actual fun <T : Any> T.ensureNeverFrozen(): T = this
+
+actual val Any?.isFrozen: Boolean get() = false
