+++ utils/src/nativeCommonMain/kotlin/com/badoo/reaktive/utils/Freeze.kt
@@ -2,6 +2,7 @@ package com.badoo.reaktive.utils
 
 import kotlin.native.concurrent.ensureNeverFrozen
 import kotlin.native.concurrent.freeze
+import kotlin.native.concurrent.isFrozen
 
 actual fun <T> T.freeze(): T = freeze()
 
@@ -10,3 +11,5 @@ actual fun <T: Any> T.ensureNeverFrozen(): T {
 
     return this
 }
+
+actual val Any?.isFrozen: Boolean get() = isFrozen
